+++ compiler/semdata.nim
@@ -332,6 +332,13 @@ proc addPragmaComputation*(c: PContext; n: PNode) =
   if c.config.symbolFiles != disabledSf:
     addPragmaComputation(c.encoder, c.packedRepr, n)
 
+proc recordPragma*(c: PContext; n: PNode; args: varargs[string]) =
+  # xxx rename to `recordAction`, since it can record things other than pragmas
+  var recorded = newNodeI(nkReplayAction, n.info)
+  for i in 0..args.high:
+    recorded.add newStrNode(args[i], n.info)
+  addPragmaComputation(c, recorded)
+
 proc inclSym(sq: var seq[PSym], s: PSym): bool =
   for i in 0..<sq.len:
     if sq[i].id == s.id: return false